+++ lib/src/collection/maps.dart
@@ -5,14 +5,14 @@ extension Maps<K, V> on Map<K, V> {
   /// Returns `true` if this nullable map is either null or empty.
   bool get isNullOrEmpty => this == null || isEmpty;
 
-  /// Returns a list containing the results of applying the given [transform] function
-  /// to each entry in the original map.
-  List<R> mapToMutableList<R>(R Function(K k, V v) transform) {
+  /// Returns a fixed-length list containing the results of applying the given
+  /// [transform] function to each entry in the original map.
+  List<R> mapToList<R>(R Function(K k, V v) transform) {
     final destination = <R>[];
     for (final entry in entries) {
       destination.add(transform(entry.key, entry.value));
     }
-    return destination;
+    return destination.toList(growable: false);
   }
 
   /// Returns the value for the given key, or the result of the [defaultValue]
@@ -21,5 +21,5 @@ extension Maps<K, V> on Map<K, V> {
 }
 
 /// Converts [MapEntry] to [Pair] with key being first component and value being second.
-Pair<K, V> parseMapEntryAsPair<K, V>(MapEntry<K, V> entry) =>
+Pair<K, V> mapEntryToPair<K, V>(MapEntry<K, V> entry) =>
     Pair(entry.key, entry.value);