+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1+3 - 2020-03-12
+### Changed
+- Change mapToMutableList() to mapToList() in maps.dart, FIXED-LENGTH list will be returned
+- Change parseMapEntryAsPair() to mapEntryToPair() in maps.dart
+
+
 ## 0.0.1+2 - 2020-03-12
 ### Added
 - Add runCatchingAsync() for async support in result.dart